# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Bia'net Platform API — Seed Marketplace, Bartar Commodity Exchange, Admin Governance
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/seed/listings:
    post:
      operationId: createSeedListing
      tags: [seed]
      summary: Create a new produce listing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeedListingInput"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeedListing"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/seed/listings/nearby:
    get:
      operationId: getNearbySeedListings
      tags: [seed]
      summary: Get nearby produce listings using geospatial search
      security:
        - bearerAuth: []
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: radius_km
          in: query
          schema:
            type: number
            default: 50
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Nearby listings with distance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NearbyListingsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/seed/listings/{id}:
    get:
      operationId: getSeedListing
      tags: [seed]
      summary: Get a produce listing by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeedListing"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateSeedListing
      tags: [seed]
      summary: Update a produce listing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeedListingUpdate"
      responses:
        "200":
          description: Listing updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeedListing"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteSeedListing
      tags: [seed]
      summary: Delete a produce listing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Listing deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/seed/orders:
    post:
      operationId: placeSeedOrder
      tags: [seed]
      summary: Place an order (initializes escrow)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeedOrderInput"
      responses:
        "201":
          description: Order placed and escrow initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscrowTransaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/seed/orders/{id}/dispute:
    post:
      operationId: disputeSeedOrder
      tags: [seed]
      summary: Raise a dispute on an escrow transaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeInput"
      responses:
        "200":
          description: Dispute raised
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscrowTransaction"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ─── BARTAR ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - limit
        - total
        - totalPages

    # ─── AUTH SCHEMAS ──────────────────────────────────────────────────────────
...
    SeedListingInput:
      type: object
      required:
        - title
        - description
        - price
        - quantity
        - unit
        - category
        - latitude
        - longitude
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
        price:
          type: string
        currency:
          type: string
          default: NGN
        quantity:
          type: string
        unit:
          type: string
        category:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        latitude:
          type: number
        longitude:
          type: number
        state:
          type: ["string", "null"]
    SeedListingUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: string
        quantity:
          type: string
        status:
          type: string
          enum: [ACTIVE, DRAFT, SUSPENDED]
        imageUrls:
          type: array
          items:
            type: string
    SeedListing:
      type: object
      required:
        - id
        - sellerId
        - title
        - price
        - quantity
        - unit
        - category
        - latitude
        - longitude
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        sellerId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        price:
          type: string
        currency:
          type: string
        quantity:
          type: string
        unit:
          type: string
        category:
          type: string
        imageUrls:
          type: ["array", "null"]
          items:
            type: string
        latitude:
          type: number
        longitude:
          type: number
        state:
          type: ["string", "null"]
        country:
          type: string
        status:
          type: string
        viewCount:
          type: integer
        distanceKm:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NearbyListingsResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SeedListing"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    SeedOrderInput:
      type: object
      required:
        - listingId
        - quantity
      properties:
        listingId:
          type: string
          format: uuid
        quantity:
          type: string
        insurance:
          type: boolean
          default: false
...
    DisputeInput:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 20

    # ─── BARTAR SCHEMAS ────────────────────────────────────────────────────────
...
    EscrowTransaction:
      type: object
      required:
        - id
        - platform
        - buyerId
        - sellerId
        - amount
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        platform:
          type: string
        listingId:
          type: string
          format: uuid
        buyerId:
          type: string
          format: uuid
        sellerId:
          type: string
          format: uuid
        amount:
          type: string
        currency:
          type: string
        platformCommission:
          type: string
        logisticsFee:
          type: string
        insuranceFee:
          type: string
        netSellerPayout:
          type: string
        status:
          type: string
        paymentReference:
          type: ["string", "null"]
        paymentProvider:
          type: ["string", "null"]
        depositedAt:
          type: ["string", "null"]
          format: date-time
        releasedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...